Extract service token provider to a seperate file
