+++ packages/server/src/app/entities/ability/index.ts
@@ -2,3 +2,4 @@ export * from './ability.entity';
 export * from './ability.module';
 export * from './ability.repository';
 export * from './ability.service';
+export * from './providers';
