+++ packages/server/src/app/entities/ability/providers.ts
@@ -1,0 +1,6 @@
+import { ABILITY_SERVICE_TOKEN, AbilityService } from './ability.service';
+
+export const abilityServiceProvider = {
+    provide: ABILITY_SERVICE_TOKEN,
+    useExisting: AbilityService,
+};
