+++ packages/server/src/app/entities/race/providers.ts
@@ -1,0 +1,6 @@
+import { RACE_SERVICE_TOKEN, RaceService } from './race.service';
+
+export const raceServiceProvider = {
+    provide: RACE_SERVICE_TOKEN,
+    useExisting: RaceService,
+};
