+++ packages/server/src/app/entities/skill/index.ts
@@ -1,3 +1,4 @@
+export * from './providers';
 export * from './skill.entity';
 export * from './skill.module';
 export * from './skill.repository';
