+++ packages/server/src/app/entities/skill/providers.ts
@@ -1,0 +1,6 @@
+import { SKILL_SERVICE_TOKEN, SkillService } from './skill.service';
+
+export const skillServiceProvider = {
+    provide: SKILL_SERVICE_TOKEN,
+    useExisting: SkillService,
+};
